backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
email-validator==2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Security
slowapi==0.1.9
python-multipart==0.0.6

# Configuration
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.25.2

# Security Scanning
bandit==1.7.5
safety==2.3.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3